requirements.txt
# GitHub Analyzer - Dependencias

# Base de datos
psycopg2-binary==2.9.9

# GitHub API
requests==2.31.0

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# Visualización
matplotlib==3.8.2
seaborn==0.13.0

# Utilidades
python-dateutil==2.8.2